Create a place to discuss technical decisions and practices
I want it to be a place where we discuss different engineering practices and write them down into a single document so we can reuse the best of them.

Some topics could be holywar-prone and in those cases we may write down some of them with disclaimer notes about usability and concerns and boundaries of applicabilities.